added dto and model mapper

+++ src/main/java/com/chs/repo/DepartmentRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.chs.entity.Department;
 
 @Repository
-public interface DepartmentRepo extends JpaRepository<Department, Integer> {
+public interface DepartmentRepo extends JpaRepository<Department, Long> {
 
 }
